+++ src/oxl_ansible_webui/aw/settings.py
@@ -125,6 +125,11 @@ AW_DB_ENGINES = {
         'PASSWORD': get_aw_env_var_or_default('db_pwd'),
         'HOST': get_aw_env_var_or_default('db_host'),
         'PORT': get_aw_env_var_or_default('db_port'),
+        'CONN_HEALTH_CHECKS': True,
+        'CONN_MAX_AGE': 0,
+        'OPTIONS': {
+            'init_command': "SET sql_mode='STRICT_TRANS_TABLES'",
+        },
     },
     'psql': {
         'ENGINE': 'django.db.backends.postgresql',
@@ -139,9 +144,7 @@ AW_DB_ENGINES = {
 DATABASES = {'default': AW_DB_ENGINES[DB_TYPE]}
 
 if DB_TYPE == 'mysql' and get_aw_env_var('db_socket') is not None:
-    DATABASES['default']['OPTIONS'] = {
-        'unix_socket': get_aw_env_var('db_socket'),
-    }
+    DATABASES['default']['OPTIONS']['unix_socket'] = get_aw_env_var('db_socket')
 
 
 def debug_mode() -> bool: